refactor(ai): create resolvers to node-wit events #22
