+++ bot/ai/index.js
@@ -1,1 +1,7 @@
-export { default as client } from './client'
+import client from './client'
+import * as resolvers from './resolvers'
+
+export {
+  client,
+  resolvers,
+}
