+++ bot/ai/resolvers/index.js
@@ -1,0 +1,1 @@
+export { default as speechAction } from './speech-action'
